+++ src/index.cjs.ts
@@ -1,3 +1,5 @@
-import Scoper from './index'
+import Scoper, { scopeIds, scopeOwnIds } from './index'
+
+Object.assign(Scoper, { scopeIds, scopeOwnIds })
 
 module.exports = Scoper